Added ID input to remove employee form

+++ UI/JS/deleteEmployee.js
@@ -2,16 +2,18 @@ const   acceptDelete = document.querySelector("#acceptDelete"),
         cancelDelete = document.querySelector("#cancelDelete"),
         deleteForm = document.querySelector("#deleteEmployeeForm"),
         scontainer = document.querySelector("#container");
+    
 
 window.addEventListener("load", () => {
 
+    
     setTimeout(() => {
 
         const delBtns = document.querySelectorAll(".removeEmployee");
 
         var hidden = true;
         
-        delBtns.forEach(btn => {
+        delBtns.forEach((btn) => {
             btn.addEventListener("click", () => {
         
                 if(hidden) {        
@@ -32,16 +34,44 @@ window.addEventListener("load", () => {
             }
         });
         
-        acceptDelete.addEventListener("click", () => {
-        
-            axios({
-                method: "DELETE",
-                url: "http://localhost:3000/employee",
-                data: {
+        acceptDelete.addEventListener("click", (e) => {
+
+            if(localStorage.getItem("token")) {
+
+                delBtns.forEach((btn, index) => {
+                    /* btnId = btn.parentNode.parentNode.children[0].innerHTML; */
+                    let btnId = index;
+                    console.log(btnId)
                     
-                }
-            });
-        
+                    userId = delBtns[btnId].parentNode.parentNode.children[0].innerHTML;
+                    console.log(userId);
+                });
+
+                
+    
+                axios({
+                    method: "DELETE",
+                    url: "http://localhost:3000/employee/" + userId,
+                    data: {
+                        idEmpleado: userId
+                    },
+                    headers: {
+                        'Authorization': "bearer " + localStorage.getItem("token")
+                    }
+                })
+                .then(response => {
+                    console.log(response);
+                })
+                .catch(error => {
+                    console.log(error);
+                });
+
+                e.preventDefault()
+            } else {
+                window.location.href = "index.html";
+            }
+            
+            e.preventDefault();
         });
 
     }, 500);
